# openapi.yaml
openapi: 3.0.0
info:
  title: 'GeoService API'
  version: 2.2.0
servers:
  -
    url: 'http://127.0.0.1:8085'
    description: 'Local environment'
paths:
...
  '/api/v1/regions/{id}':
    put:
      tags:
        - Regions
      operationId: put_app_region_api_v1_region_update
      parameters:
        -
          name: id
          in: path
          description: 'Region uuid'
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRegionActionRequest'
      responses:
        '200':
          description: 'Region updated successfully.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegionResponse'
    delete:
      tags:
        - Regions
      operationId: delete_app_region_api_v1_region_delete
      parameters:
        -
          name: id
          in: path
          description: 'Region uuid'
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Region deleted successfully.'
  /api/v1/subregions:
    get:
      tags:
        - SubRegions
      operationId: get_app_subregion_api_v1_subregion_list
      parameters:
        -
          name: limit
          in: query
          description: 'Default 500'
          required: false
          content:
            application/json:
              schema:
                type: integer
        -
          name: offset
          in: query
          description: 'Default 0'
          required: false
          content:
            application/json:
              schema:
                type: integer
        -
          name: title
          in: query
          description: 'SubRegion title'
          required: false
          content:
            application/json:
              schema:
                type: string
      responses:
        '200':
          description: 'List of subregions.'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SubRegionResponse'
    post:
      tags:
        - SubRegions
      operationId: post_app_subregion_api_v1_subregion_create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubRegionActionRequest'
      responses:
        '200':
          description: 'Nationality created successfully.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubRegionResponse'
  '/api/v1/subregions/{id}':
    put:
      tags:
        - SubRegions
      operationId: put_app_subregion_api_v1_subregion_update
      parameters:
        -
          name: id
          in: path
          description: 'SubRegion uuid'
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSubRegionActionRequest'
      responses:
        '200':
          description: 'SubRegion updated successfully.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubRegionResponse'
    delete:
      tags:
        - SubRegions
      operationId: delete_app_subregion_api_v1_subregion_delete
      parameters:
        -
          name: id
          in: path
          description: 'SubRegion uuid'
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'SubRegion deleted successfully.'
  /api/v1/timezones:
    get:
      tags:
        - Timezones
      operationId: get_app_timezone_api_v1_timezone_list
      parameters:
        -
          name: limit
          in: query
          description: 'Default 500'
          required: false
          content:
            application/json:
              schema:
                type: integer
        -
          name: offset
          in: query
          description: 'Default 0'
          required: false
          content:
            application/json:
              schema:
                type: integer
        -
          name: title
          in: query
          description: 'Timezone title'
          required: false
          content:
            application/json:
              schema:
                type: string
        -
          name: code
          in: query
          description: 'Timezone code'
          required: false
          content:
            application/json:
              schema:
                type: string
        -
          name: utc
          in: query
          description: 'Timezone UTC'
          required: false
          content:
            application/json:
              schema:
                type: string
      responses:
        '200':
          description: 'List of timezones.'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TimezoneResponse'
    post:
      tags:
        - Timezones
      operationId: post_app_timezone_api_v1_timezone_create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTimezoneActionRequest'
      responses:
        '200':
          description: 'Nationality created successfully.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimezoneResponse'
  '/api/v1/timezones/{id}':
    put:
      tags:
        - Timezones
      operationId: put_app_timezone_api_v1_timezone_update
      parameters:
        -
          name: id
          in: path
          description: 'Timezone uuid'
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTimezoneActionRequest'
      responses:
        '200':
          description: 'Timezone updated successfully.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimezoneResponse'
    delete:
      tags:
        - Timezones
      operationId: delete_app_timezone_api_v1_timezone_delete
      parameters:
        -
          name: id
          in: path
          description: 'Timezone uuid'
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Timezone deleted successfully.'
...
components:
  schemas:
...
    RegionResponse:
      required:
        - title
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
      type: object
...
    UpdateRegionActionRequest:
      required:
        - title
      properties:
        title:
          type: string
      type: object
    SubRegionResponse:
      required:
        - title
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        region:
          nullable: true
          oneOf:
            -
              $ref: '#/components/schemas/RegionResponse'
      type: object
    CreateSubRegionActionRequest:
      required:
        - title
        - regionTitle
      properties:
        title:
          type: string
          maxLength: 150
          minLength: 1
        regionTitle:
          type: string
          maxLength: 150
          minLength: 1
      type: object
    UpdateSubRegionActionRequest:
      required:
        - title
      properties:
        title:
          type: string
        regionTitle:
          type: string
          nullable: true
      type: object
    TimezoneResponse:
      required:
        - title
        - code
        - utc
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        code:
          type: string
        utc:
          type: string
      type: object
    CreateTimezoneActionRequest:
      required:
        - title
        - code
        - utc
      properties:
        title:
          type: string
          maxLength: 150
          minLength: 1
        code:
          type: string
          maxLength: 150
          minLength: 1
        utc:
          type: string
          maxLength: 50
          minLength: 1
      type: object
    UpdateTimezoneActionRequest:
      required:
        - title
        - code
        - utc
      properties:
        title:
          type: string
        code:
          type: string
        utc:
          type: string
      type: object
...
  securitySchemes:
    Bearer:
      type: http
      bearerFormat: JWT
      scheme: bearer
security:
  -
    Bearer: []